k1max-ws-bridge-github/k1max-ws-bridge/k1max-ws-bridge/requirements.txt
websockets==12.0
paho-mqtt==1.6.1
jsonpath-ng==1.6.1
